CI, 解决证书导入问题，脚本完善
Close #23

+++ CI/gitlab.sh
@@ -1,4 +1,7 @@
 #! /bin/zsh
+# 适用于 GitLab 的 iOS 项目打包脚本
+# Copyright © 2020 BB9z.
+# https://github.com/BB9z/iOS-Project-Template 
 
 set -euo pipefail
 echo "$PWD"
@@ -24,13 +27,17 @@ logSection() {
 logSection "环境检查"
 export LANG=en_US.UTF-8
 
-# 分支名
-readonly CI_COMMIT_MESSAGE=${CI_COMMIT_MESSAGE:="$(logWarning 'CI_COMMIT_REF_NAME 未设置，可能非 GitLab CI 环境')"}
+readonly CI_COMMIT_MESSAGE=${CI_COMMIT_MESSAGE:="$(logWarning 'CI_COMMIT_MESSAGE 未设置，可能非 GitLab CI 环境')"}
 logInfo "CI_COMMIT_MESSAGE = $CI_COMMIT_MESSAGE"
 isVerbose=false
 if [[ "$CI_COMMIT_MESSAGE" = *"[ci verbose]"* ]]; then
     logWarning "verbose 已激活"
     isVerbose=true
+fi
+isCleanEnabled=false
+if [[ "$CI_COMMIT_MESSAGE" = *"[ci clean]"* ]]; then
+    logWarning "清理已激活"
+    isCleanEnabled=true
 fi
 
 readonly XC_BUILD_SCHEME="${XC_BUILD_SCHEME:? 必须设置}"
@@ -40,11 +47,11 @@ logInfo "XC_BUILD_WORKSPACE = $XC_BUILD_WORKSPACE"
 readonly XC_BUILD_CONFIGURATION="${XC_BUILD_CONFIGURATION:="$(logWarning 'XC_BUILD_CONFIGURATION 未设置，默认 Release')Release"}"
 logInfo "XC_BUILD_CONFIGURATION = $XC_BUILD_CONFIGURATION"
 
-readonly XC_PROVISIONING_ID=${XC_PROVISIONING_ID:="$(logWarning 'XC_PROVISIONING_ID 未设置')"}
+readonly XC_PROVISIONING_ID=${XC_PROVISIONING_ID:="$(logWarning 'XC_PROVISIONING_ID 未设置，将尝试自动签名')"}
 if [ -n "$XC_PROVISIONING_ID" ]; then
     logInfo "XC_PROVISIONING_ID = $XC_PROVISIONING_ID"
 fi
-readonly XC_CODE_SIGN_IDENTITY=${XC_CODE_SIGN_IDENTITY:="$(logWarning 'XC_CODE_SIGN_IDENTITY 未设置')"}
+readonly XC_CODE_SIGN_IDENTITY=${XC_CODE_SIGN_IDENTITY:="$(logWarning 'XC_CODE_SIGN_IDENTITY 未设置，将尝试自动签名')"}
 if [ -n "$XC_CODE_SIGN_IDENTITY" ]; then
     logInfo "XC_CODE_SIGN_IDENTITY = $XC_CODE_SIGN_IDENTITY"
 fi
@@ -54,7 +61,11 @@ readonly XC_IMPORT_CERTIFICATE_PATH=${XC_IMPORT_CERTIFICATE_PATH:=""}
 logInfo "XC_IMPORT_CERTIFICATE_PATH = $XC_IMPORT_CERTIFICATE_PATH"
 readonly XC_IMPORT_CERTIFICATE_PASSWORD=${XC_IMPORT_CERTIFICATE_PASSWORD:=""}
 
-FIR_UPLOAD_TOKEN=${FIR_UPLOAD_TOKEN:="$(logWarning 'FIR_UPLOAD_TOKEN 未设置')"}
+# keychian 的设置可以从外部导入，但默认不推荐修改
+readonly KC_NAME=${KC_NAME:="CIBuilder"}
+readonly KC_PASSWORD=${KC_PASSWORD:="ci"}
+
+FIR_UPLOAD_TOKEN=${FIR_UPLOAD_TOKEN:="$(logWarning 'FIR_UPLOAD_TOKEN 未设置，将跳过 fir.im 上传')"}
 if [ -n "$FIR_UPLOAD_TOKEN" ]; then
     logInfo "FIR_UPLOAD_TOKEN 将从 $FIR_UPLOAD_TOKEN 变量读取"
     FIR_UPLOAD_TOKEN=$(eval echo -e "\$$FIR_UPLOAD_TOKEN")
@@ -69,28 +80,43 @@ readonly EXPORT_IPA_PATH="$EXPORT_DIRECTORY_PATH/$XC_BUILD_SCHEME.ipa"
 xcodebuild -version
 
 logSection "配置更新"
-isNeedsPodInstall=false
-diff "Podfile.lock" "Pods/Manifest.lock" >/dev/null || {
-    isNeedsPodInstall=true
-}
-if $isNeedsPodInstall; then
-    logInfo "执行 CocoaPods"
-    local installOption=$(( $isVerbose && echo "" ) || echo "--silent" )
-    pod install $installOption || {
-        logError "pod install 失败，尝试更新 repo"
-        pod install --repo-update
+if [ -f "Podfile" ]; then
+    local isNeedsPodInstall=false
+    diff "Podfile.lock" "Pods/Manifest.lock" >/dev/null || {
+        isNeedsPodInstall=true
     }
+    if $isNeedsPodInstall; then
+        logInfo "执行 CocoaPods"
+        local installOption=$(( $isVerbose && echo "" ) || echo "--silent" )
+        pod install $installOption || {
+            logError "pod install 失败，尝试更新 repo"
+            pod install --repo-update
+        }
+    else
+        logInfo "Podfile 未变化，跳过 CocoaPods 安装"
+    fi
 else
-    logInfo "Podfile 未变化，跳过 CocoaPods 安装"
+    logWarning "Podfile 不存在？跳过 CocoaPods 安装"
 fi
 
+isCIKeycahinCreated=false
 if [ -n "$XC_IMPORT_CERTIFICATE_PATH" ]; then
     if [ -z "$XC_IMPORT_CERTIFICATE_PASSWORD" ]; then
         logError "安装证书已指定，但是密码未设置"
         exit 1
     fi
+    logInfo "创建临时 keychain"
+    security create-keychain -p "$KC_PASSWORD" "$KC_NAME.keychain" || {
+        logError "创建临时 keychain 失败，$KC_NAME 可能已存在，你可以通过设置 KC_NAME 环境变量换一个"
+        exit 1
+    }
+    isCIKeycahinCreated=true
+    security list-keychains -d user -s "$KC_NAME.keychain" $(security list-keychains -d user | sed s/\"//g)
+    security set-keychain-settings "$KC_NAME.keychain"
+    security unlock-keychain -p "$KC_PASSWORD" "$KC_NAME.keychain"
     logInfo "导入签名证书"
-    security import "$XC_IMPORT_CERTIFICATE_PATH" -P "$XC_IMPORT_CERTIFICATE_PASSWORD" -T /usr/bin/codesign
+    security import "$XC_IMPORT_CERTIFICATE_PATH" -k "$KC_NAME.keychain" -P "$XC_IMPORT_CERTIFICATE_PASSWORD" -T "/usr/bin/codesign"
+    security set-key-partition-list -S apple-tool:,apple: -s -k "$KC_PASSWORD" "$KC_NAME.keychain"
 fi
 
 if [ -n "$XC_IMPORT_PROVISIONING_PATH" ]; then
@@ -98,14 +124,14 @@ if [ -n "$XC_IMPORT_PROVISIONING_PATH" ]; then
     open "$XC_IMPORT_PROVISIONING_PATH"
 fi
 
+
 logSection "项目构建"
-if [[ "$CI_COMMIT_MESSAGE" = *"[ci clean]"* ]]; then
-    logWarning "提交信息指定项目清理"
-    xcodebuild clean -workspace "$XC_BUILD_WORKSPACE" -scheme "$XC_BUILD_SCHEME" | xcpretty
+xcprettyOptions=$(( $isVerbose && echo "" ) || echo "-t" )
+
+if $isCleanEnabled; then
+    xcodebuild clean -workspace "$XC_BUILD_WORKSPACE" -scheme "$XC_BUILD_SCHEME" | xcpretty $xcprettyOptions
 fi
 
-buildAddtionalOptions=""
-xcprettyOptions=$(( $isVerbose && echo "" ) || echo "-t" )
 if [ -n "$XC_PROVISIONING_ID" ] && [ -n "$XC_CODE_SIGN_IDENTITY" ]; then
     logInfo "指定签名，开始编译..."
     xcodebuild archive -archivePath "$ARCHIVE_PATH" \
@@ -114,7 +140,7 @@ if [ -n "$XC_PROVISIONING_ID" ] && [ -n "$XC_CODE_SIGN_IDENTITY" ]; then
         CODE_SIGN_STYLE="Manual" PROVISIONING_PROFILE_SPECIFIER="$XC_PROVISIONING_ID" CODE_SIGN_IDENTITY="$XC_CODE_SIGN_IDENTITY" |
         xcpretty $xcprettyOptions
 else
-    logInfo "XC_PROVISIONING_ID 或 XC_CODE_SIGN_IDENTITY 未设置，尝试自动签名，开始编译..."
+    logInfo "尝试自动签名，开始编译..."
     xcodebuild archive -archivePath "$ARCHIVE_PATH" \
         -workspace "$XC_BUILD_WORKSPACE" -scheme "$XC_BUILD_SCHEME" \
         -configuration "$XC_BUILD_CONFIGURATION" -destination generic/platform=iOS |
@@ -123,6 +149,11 @@ fi
 
 logSection "项目打包"
 xcodebuild -exportArchive -archivePath "$ARCHIVE_PATH" -exportPath "$EXPORT_DIRECTORY_PATH" -exportOptionsPlist "$EXPORT_OPTIONS_PLIST" | xcpretty $xcprettyOptions
+
+if $isCIKeycahinCreated; then
+    logInfo "清理临时 keychain"
+    security delete-keychain "$KC_NAME.keychain"
+fi
 
 dSYMsCount=$(find "$ARCHIVE_PATH/dSYMs" -d 1 -name "*.dSYM" | wc -l)
 if [ $dSYMsCount -ge 1 ]; then
@@ -142,3 +173,6 @@ if [ -n "$FIR_UPLOAD_TOKEN" ]; then
 else
     logInfo "跳过上传"
 fi
+
+echo ""
+logInfo "🎉 一切正常"
